a reader for the image that is not us: icheck and dcheck
Everything tests/mkfs asserted before this asks whether the image matches
what THIS PORT believes a V8 filesystem is -- and the belief and the
bytes come from the same place, so a shared misunderstanding satisfies
both.  icheck and dcheck are 1985 code that knows nothing about mkfs.

	files      2 (r=1,d=1,b=0,c=0,l=0)
	used       1 (i=0,ii=0,iii=0,d=1)
	free    1917
	missing    0

`free' is walked out of the free list block by block through the fblk
chain; s_tfree is a counter mkfs kept while writing.  Two independent
computations, and used + free + s_isize is exactly s_fsize with nothing
missing -- if NICFREE, the fblk layout or daddr_t's width were wrong by
one the walk would end early and the arithmetic would not close.  The
suite asserts the arithmetic rather than 1917.

They also gave l3tol its first caller in this port.  ltol3 writes those
3-byte addresses, l3tol reads them, both were patched and then re-patched
when daddr_t narrowed, and nothing had ever read one back.

THE CORRUPTION CASES ARE THE HALF THAT PROVES ANYTHING.  Three, each with
a different correct diagnosis: an orphaned block reads as `missing 1', an
out-of-range address is named with its inode, and a wrong link count is
seen by dcheck while icheck correctly stays silent.  A checker that
answered everything would be the more suspicious result.

ONE PATCHED LINE between the two programs, and it was the third instance
of narrowing a type at the seam and meeting something that already
encoded the old width -- ltol3 and l3tol were the others.  icheck's
alloc() walks the BITFS bitmap with a raw `long *'.  It is the only one
of the three the compiler diagnosed, because the other two do their
arithmetic through a char * where the type is inert.  The rule that comes
out: sweep for raw pointers first, since subscripting follows the element
type on its own.

-DDIRSIZ=14 is now a property of the GROUP rather than a flag on mkfs,
because dcheck parses the records mkfs writes.  A forgotten flag there is
worse than in mkfs: it would report a healthy filesystem as healthy and
only start lying once a directory held more than one entry per 256 bytes.

Two tests found their own defects on the way.  tests/jail asked whether
`ls /etc | head -1' was `fstab' -- a value where a relation was meant,
true only while nothing installed into /etc sorted before it.  Adding
dcheck turned it red with nothing about the jail changed; it now compares
the whole listing against the rootfs's own /etc, which says strictly more
and cannot drift.  And the first deps cases for these rules asserted
header -> binary, which passes for the wrong reason.

Seventeen suites, 1018 cases.  All six earlier fixes plus these are
mutation-verified.

+++ v8/src/cmd/icheck.c
@@ -368,7 +368,23 @@ alloc()
 		struct	fblk fb;
 	} buf;
 	register int i, j;
-	register long *p;
+	/*
+	 * PORT: upstream `register long *p'.  It walks sblock.s_bfree, the
+	 * BITFS free-block bitmap, which <sys/filsys.h> now declares `int'
+	 * because a disk record's fields are the widths a VAX gave them and
+	 * NOLONG made `long' 32 bits there.  Upstream's own comment eleven
+	 * lines down says "BITS PER LONG" beside a loop to 32, so the type this
+	 * wants is whatever is 32 bits wide, and here that is int.
+	 *
+	 * THIRD INSTANCE of narrowing a type at the seam and finding something
+	 * that had already encoded the old width -- ltol3.c and l3tol.c were
+	 * the other two, and this is the only one the compiler said anything
+	 * about.  The others index by subscript, which follows the element type
+	 * on its own; a raw pointer does not.  Swept: `grep -rnE "long[ \t]*\*"'
+	 * against every reader of s_free, s_bfree, i_addr and di_addr finds
+	 * this line and nothing else.
+	 */
+	register int *p;
 
 	sblock.s_tfree--;
 	if (BITFS(dev)) {
